requirements.txt
Flask==2.3.2
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.23.5  # Changed from 1.24.3 to fit TensorFlow's requirement